+++ flake.nix
@@ -15,7 +15,7 @@
       let
         pkgs = import nixpkgs { inherit system; };
 
-        version = "1.4.0";
+        version = "1.5.0";
         commitOrDirty = if self ? rev then self.rev else "dirty";
 
         breezyd-pkg = pkgs.buildGoModule {